requirements.txt
anthropic>=0.49.0
rich>=13.7.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
